Added Edit and Create funcitonallity for Events.
Users can now Create a new event which is saved to the database. They can also edit existing events.

Additional changes: Set both 'GetEvents' and 'GetUserEvents' to order by Event_Date. 'GetEvents' is ordered within the model, however 'GetUserEvents' is done in the view as the data being displayed in the view is coming from a referenced table.

Next step: To configure the function to add and remove an event to a specific users event history.