Fixed Authentication & Half implemented user event history.
Users can now add an event to their event history, however the fields are having to be inputted manually. This needs to be changed to ensure that no false data is inputted, and the input boxes on the AddToUser view should be automatically filled with the correct data from @htmlactionlink on previous page.