Added functionality for removing the 'Add To My Events' Link on GetEvents.
Implemented a solution to remove 'Add To My Events' from showing on all events regardless of whether the event has already been added to the users profile or not.

This solution makes use of a viewbag populated by the GetUserEvents action using the current logged in user. Within the 'GetEvents' view, an additional if statement was added to check this viewbag to see if it contains the Event_ID for each event in 'tbl_events'. If it exists, 'TEST' is shown, if it doens't, 'Add To My Events' is shown.

'TEST' is to be replaced by the remove function still to be implemented.