Implemented funciton to return view of # of times a user has seen a dj.
Implemented a funciton which uses an SQL statement to return how many times a user has seen each DJ in their history. This uses a DataTable which is converted into a List, and displayed on a view. This view needs to be worked on to be shown in the standard 'SeenArtists' indexed view.