Implemented new Venue Controller & CRUD Actions.
Implemented a new controller (Venue) to allow improved handling of an events location as it provides the scope to select a venue from the database when creating/editing an event. This controller refers to a new DB table which holds a specific venue's details (Name, Address  etc.). CRUD Actions have been implemented for this table.

Additionally, a new database table has been added containing a list of countries, which is used to fill the Venue's country field.