Controller Action Changes & View Updates.
Function of 'AdminController' is to now handle all Create, Edit & Delete actions for Events, Artists, Venues and Countries (this will be locked to only users with admin permissions (to be added further ahead)). This change has meant that existing functions within the other controllers have been moved to the AdminController, with the views moving to the respective 'Admin' view folder.

The original Index views have been moved to the Admin section, and a new Event Index view has been created for normal users.

Normal User views have yet to be created for Venues and Artists.