UserEvents View added with filtering. Change to DB structure.
Implemented a new table 'tbl_cities', which is a table which contains names of individual cities in each country. This will improve venue creation in the future.

Also, made steps on creating a 'GetEvents' view for normal users. This view makes use of scripts to filter the data (Events in the database) in a partial view grid, based on the selected: Country; City; Venue.