Multiple changes made to improve implementation of SeenArtistCount.
Extended SeenArtistCount Model to include a nullable int which is used to store the File_ID of each Artists corresponding ‘Profile Image'. This has been achieved by modifying the SQL Command used within the GetUsersArtistCount Action, which now returns a list of a users most seen artists and their corresponding File_ID (If there is no File Found for that artists within tbl_artistImages, Null is returned). This enables artists to be displayed in the view who don’t have an image associated with them, who now the ‘default-profile.jpeg’ is displayed for.

The GetUsersArtistCount view has been updated to look more aesthetically pleasing, with the users top 3 seen artists being displayed in seperate containers at the top of the page. The rest of the list is displayed below. Styling is applied differently to containers for artists ranked 1,2 and 3, with then a standard styling being applied to rank 4 onwards.

GetSeenArtistHistory View has been updated to match the styling of EventIndex views.

Created ContentResult within ApplicationController to return TrackerEntities ConnectionString from Web.config using this solution (https://stackoverflow.com/a/45030946/12764653). This now means that instead of SQL Connections requiring the Connection String to be written each time a method that requires it is written, they can instead call this ContentResult. This will hopefully enforce these Connections to work upon modification of the Connection String (E.g. Windows Auth -> SQL Auth).